diagonal walking issues are twofold: player keeps moving when keys are released: somthing in KBcontroller, diagonal movement not always 45 deg: sothing with normalisation

+++ src/Vector2D.cpp
@@ -20,7 +20,7 @@ Vector2D &Vector2D::Zero()
 
 Vector2D &Vector2D::Normalize()
 {
-    double L = sqrt(this->x * this->x + this->y * this->y);
+    float L = sqrtf(this->x * this->x + this->y * this->y);
     this->x /= L;
     this->y /= L;
